refactor(lowlevel): rework into proper library code
