+++ lowlevel/doc.go
@@ -1,0 +1,5 @@
+/*
+Package lowlevel 包含与企业微信服务端沟通相关的原语。
+*/
+
+package lowlevel
